添加单元测试，并修复BUG。

+++ hutool-bloomFilter/src/main/java/cn/hutool/bloomfilter/bitMap/IntBitMap.java
@@ -7,8 +7,8 @@ package cn.hutool.bloomfilter.bitmap;
  */
 public interface IntBitMap extends BitMap {
 
-	int getMin();
+	long getMin();
 
-	int getMax();
+	long getMax();
 
 }
